Recipe selection algorithm and chain fusion
fixed #16
...wait that was fast

Algorithm browses through every weapon combination resulting in a recipe.
If the ingredient as input are equal to that combinaation, their priority is updated.
At the end of the browsing the highest found priority recipe is returned.

Also chain fusion!! 3 >= weapons in table of crafting fuse them 2 by 2
with a 35 tics delay between each.

"none" < "---"