Weapon icon ; recipe tips (non LMS only)
A random recipe is given on map start. If the recipes set is too big to compute the game picks
a selection of weapons (about 13 max) and find the recipes with them.
The weapon picker algorithm agilely morph to reduce computer cost .

Also it display wep icons for that recipe