Algorithm improvement, code restruct
Now uses ints instead of strings when possible. "null" string is "?" and "null"
number is -1. Idea to implement the algorithm : keywords reduced to 3,
only investigating the corresponding ND range of the cookbook,
breaking research when a "?" keyword is found.
However placeholders "s" labels are being kept in practice so we mod
considering a potentiall all weapons having 3 keywords. Worst case scenario.

With these changes we can compute up to about 33 weapon combinations.